Reorganize directory structure
- Move the modules under the `sync_base` module to the `sync` and `common` modules.
