+++ src/sync.rs
@@ -1,12 +1,25 @@
 //! Provides thread-safe, concurrent cache implementations.
 
+mod base_cache;
 mod builder;
 mod cache;
 mod entry_selector;
+mod invalidator;
+mod key_lock;
 mod segment;
 mod value_initializer;
 
-pub use crate::sync_base::{iter::Iter, PredicateId};
+/// The type of the unique ID to identify a predicate used by
+/// [`Cache::invalidate_entries_if`][invalidate-if] method.
+///
+/// A `PredicateId` is a `String` of UUID (version 4).
+///
+/// [invalidate-if]: ./struct.Cache.html#method.invalidate_entries_if
+pub type PredicateId = String;
+
+pub(crate) type PredicateIdStr<'a> = &'a str;
+
+pub use crate::common::iter::Iter;
 pub use {
     builder::CacheBuilder,
     cache::Cache,
